Adds initial javadoc comment blocks to clean up warnings

+++ lib/src/main/java/ninja/codingsolutions/solaredgeapiclient/models/EnergyData.java
@@ -2,7 +2,10 @@ package ninja.codingsolutions.solaredgeapiclient.models;
 
 import lombok.Data;
 
+/**
+ * Energy data in kWh
+ */
 @Data
 public class EnergyData {
-    private int energy;
+    private double energy;
 }
